Transmit pin status back to requester.

+++ globalVar.h
@@ -1,7 +1,8 @@
 #ifndef GLOBALVAR_H
 #define GLOBALVAR_H
 
+#include <QCoreApplication>
 
-extern int deviceStatusArray[2];
+extern qint8 deviceStatusArray[2];
 
 #endif // GLOBALVAR_H
